Move test fragment and layout resource to fixture library module for each binding library module for better IDE support.
