--- flowbinding-material/src/androidTest/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="reactivecircus.flowbinding.material.test" />
