+++ flowbinding-material/src/androidTest/java/reactivecircus/flowbinding/material/BottomSheetBehaviorSlidedFlowTest.kt
@@ -7,6 +7,7 @@ import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 import reactivecircus.blueprint.testing.action.swipeDownOnView
 import reactivecircus.blueprint.testing.action.swipeUpOnView
+import reactivecircus.flowbinding.material.fixtures.MaterialFragment
 import reactivecircus.flowbinding.material.test.R
 import reactivecircus.flowbinding.testing.FlowRecorder
 import reactivecircus.flowbinding.testing.launchTest
@@ -20,7 +21,6 @@ class BottomSheetBehaviorSlidedFlowTest {
         launchTest<MaterialFragment> {
             val recorder = FlowRecorder<Float>(testScope)
             val bottomSheet = getViewById<View>(R.id.bottomSheetLayout)
-            val behavior = BottomSheetBehavior.from(bottomSheet)
             bottomSheet.bottomSheetSlides().recordWith(recorder)
 
             recorder.assertNoMoreValues()
