+++ flowbinding-material/src/androidTest/java/reactivecircus/flowbinding/material/SnackbarDismissedFlowTest.kt
@@ -9,6 +9,7 @@ import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 import reactivecircus.blueprint.testing.action.clickSnackbarActionButton
 import reactivecircus.blueprint.testing.asViewAction
+import reactivecircus.flowbinding.material.fixtures.MaterialFragment
 import reactivecircus.flowbinding.material.test.R
 import reactivecircus.flowbinding.testing.FlowRecorder
 import reactivecircus.flowbinding.testing.launchTest
