+++ flowbinding-material/src/androidTest/java/reactivecircus/flowbinding/material/SwipeDismissBehaviorDismissedFlowTest.kt
@@ -7,6 +7,7 @@ import com.google.android.material.behavior.SwipeDismissBehavior
 import com.google.android.material.card.MaterialCardView
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
+import reactivecircus.flowbinding.material.fixtures.MaterialFragment
 import reactivecircus.flowbinding.material.test.R
 import reactivecircus.flowbinding.testing.FlowRecorder
 import reactivecircus.flowbinding.testing.hardSwipeLeft
