+++ flowbinding-viewpager2/fixtures/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="reactivecircus.flowbinding.viewpager2.fixtures" />
