+++ flowbinding-viewpager2/fixtures/src/main/res/layout/fragment_viewpager2.xml
