+++ flowbinding-viewpager2/fixtures/src/main/res/layout/item_viewpager.xml
