--- flowbinding-viewpager2/src/androidTest/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="reactivecircus.flowbinding.viewpager2.test" />
